+++ src/main/scala/io/github/karlhigley/lexrank/LexRank.scala
@@ -5,21 +5,21 @@ import scala.math.max
 import org.apache.spark.rdd.RDD
 import org.apache.spark.graphx._
 
-case class SentenceComparison(id1: Long, id2: Long, similarity: Double)
+case class SentenceComparison(id1: Long, id2: Long, similarity: Float)
 
-class LexRank(graph: Graph[String, Double]) extends Serializable {
-  def score(cutoff: Double = 0.8, convergence: Double = 0.001) : VertexRDD[Double] = {
+class LexRank(graph: Graph[String, Float]) extends Serializable {
+  def score(cutoff: Float = 0.8f, convergence: Float = 0.001f) : VertexRDD[Double] = {
     filterBoilerplate(graph, cutoff).pageRank(convergence).vertices
   }
 
-  private def filterBoilerplate(graph: Graph[String, Double], cutoff: Double): Graph[String, Double] = {
-    val maxSimilarityVertices = Graph.fromEdges(graph.edges, 0).aggregateMessages[Double](
+  private def filterBoilerplate(graph: Graph[String, Float], cutoff: Float): Graph[String, Float] = {
+    val maxSimilarityVertices = Graph.fromEdges(graph.edges, 0).aggregateMessages[Float](
       sendMsg       = triplet => triplet.sendToDst(triplet.attr),
       mergeMsg      = (a, b) => max(a,b),
       tripletFields = TripletFields.EdgeOnly
     )
 
-    val maxSimilarityGraph: Graph[Double, Double] = Graph(maxSimilarityVertices, graph.edges, 1.0)
+    val maxSimilarityGraph: Graph[Float, Float] = Graph(maxSimilarityVertices, graph.edges, 1.0f)
     val filteredGraph = maxSimilarityGraph.subgraph(vpred = (id, attr) => attr < cutoff)
 
     graph.mask(filteredGraph).subgraph(
